+++ installer/build-installer.ps1
@@ -8,7 +8,7 @@ param(
     [ValidateSet('Debug', 'Release')]
     [string]$Configuration = 'Release',
     
-    [string]$Version = '2.31.1.0',
+    [string]$Version = '2.31.2.0',
 
     [string]$FfmpegArchivePath
 )